# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "API Documentation",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8001",
            "description": "Documentation API Server"
        }
    ],
    "paths": {
...
        "/api/v1/shop/categories": {
            "get": {
                "tags": [
                    "Category for shops"
                ],
                "summary": "get all categories. => for shopkeeper and system admin",
                "description": "get all categories",
                "operationId": "get categories",
                "parameters": [
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Tags to sort by",
                        "required": false,
                        "style": "form",
                        "schema": {
                            "type": "string"
                        },
                        "example": "id"
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "Tags to order by 'asc|desc'",
                        "required": false,
                        "style": "form",
                        "schema": {
                            "type": "string"
                        },
                        "example": "asc"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Tags to paginate for page number",
                        "required": false,
                        "style": "form",
                        "schema": {
                            "type": "number"
                        },
                        "example": "1"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Tags for number of limit item in paginate result",
                        "required": false,
                        "style": "form",
                        "schema": {
                            "type": "number"
                        },
                        "example": "10"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "data of result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "number",
                                            "example": "1"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "category_name": {
                                                        "type": "string",
                                                        "example": "category_name"
                                                    },
                                                    "id": {
                                                        "type": "integer",
                                                        "example": "11"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "first_page_url": {
                                            "type": "string",
                                            "example": "http://127.0.0.1:8001/api/v1/shop/categories?page=1"
                                        },
                                        "last_page_url": {
                                            "type": "string",
                                            "example": "http://127.0.0.1:8001/api/v1/shop/categories?page=5"
                                        },
                                        "last_page": {
                                            "type": "number",
                                            "example": "5"
                                        },
                                        "from": {
                                            "type": "number",
                                            "example": "1"
                                        },
                                        "links": {
                                            "example": "{'url': null,'label': '&laquo; Previous','active': false},{'url': 'http://127.0.0.1:8001/api/v1/shop/categories?page=1','label': '1','active': true},{'url': null,'label': 'Next &raquo;','active': false}"
                                        },
                                        "next_page_url": {
                                            "type": "string",
                                            "example": "http://127.0.0.1:8001/api/v1/shop/categories?page=2"
                                        },
                                        "prev_page_url": {
                                            "type": "string",
                                            "example": "null"
                                        },
                                        "path": {
                                            "type": "string",
                                            "example": "http://127.0.0.1:8001/api/v1/shop/categories"
                                        },
                                        "per_page": {
                                            "type": "string",
                                            "example": "'10'"
                                        },
                                        "to": {
                                            "type": "number",
                                            "example": "3"
                                        },
                                        "total": {
                                            "type": "number",
                                            "example": "13"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "data of result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "number",
                                            "example": "1"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "category_name": {
                                                        "type": "string",
                                                        "example": "category_name"
                                                    },
                                                    "id": {
                                                        "type": "integer",
                                                        "example": "11"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "first_page_url": {
                                            "type": "string",
                                            "example": "http://127.0.0.1:8001/api/v1/shop/categories?page=1"
                                        },
                                        "last_page_url": {
                                            "type": "string",
                                            "example": "http://127.0.0.1:8001/api/v1/shop/categories?page=5"
                                        },
                                        "last_page": {
                                            "type": "number",
                                            "example": "5"
                                        },
                                        "from": {
                                            "type": "number",
                                            "example": "1"
                                        },
                                        "links": {
                                            "example": "{'url': null,'label': '&laquo; Previous','active': false},{'url': 'http://127.0.0.1:8001/api/v1/shop/categories?page=1','label': '1','active': true},{'url': null,'label': 'Next &raquo;','active': false}"
                                        },
                                        "next_page_url": {
                                            "type": "string",
                                            "example": "http://127.0.0.1:8001/api/v1/shop/categories?page=2"
                                        },
                                        "prev_page_url": {
                                            "type": "string",
                                            "example": "null"
                                        },
                                        "path": {
                                            "type": "string",
                                            "example": "http://127.0.0.1:8001/api/v1/shop/categories"
                                        },
                                        "per_page": {
                                            "type": "string",
                                            "example": "'10'"
                                        },
                                        "to": {
                                            "type": "number",
                                            "example": "3"
                                        },
                                        "total": {
                                            "type": "number",
                                            "example": "13"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "405": {
                        "description": "Method Not Allowed"
                    },
                    "406": {
                        "description": "Not Acceptable"
                    },
                    "407": {
                        "description": "Proxy Authentication Required"
                    },
                    "410": {
                        "description": "Resource Gone"
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "423": {
                        "description": "Resource Locked"
                    },
                    "429": {
                        "description": "Too Many Requests"
                    },
                    "451": {
                        "description": "Unavailable For Legal Reasons"
                    },
                    "500": {
                        "description": "Internal Server"
                    },
                    "502": {
                        "description": "Bad Gateway"
                    },
                    "503": {
                        "description": "Service Unavailable"
                    },
                    "504": {
                        "description": "Gateway Timeout"
                    },
                    "505": {
                        "description": "HTTP Version Not Supported"
                    },
                    "511": {
                        "description": "Network Authentication Required"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Category for shops"
                ],
                "summary": "create new category. => only for system admin",
                "description": "create new category",
                "operationId": "newCategory",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "category_name"
                                ],
                                "properties": {
                                    "category_name": {
                                        "description": "category name",
                                        "type": "string",
                                        "format": "string",
                                        "example": "category name"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {
                                "required": [
                                    "category_name"
                                ],
                                "properties": {
                                    "category_name": {
                                        "nullable": true,
                                        "description": "category name",
                                        "type": "string",
                                        "format": "string",
                                        "default": "null",
                                        "example": "category name"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "New Category Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "category_id": {
                                            "type": "integer",
                                            "example": "12"
                                        },
                                        "category_name": {
                                            "type": "string",
                                            "example": "category_name"
                                        },
                                        "created_at": {
                                            "type": "timestamp",
                                            "example": "2021-08-05 01:19:35"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "New Category Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "category_id": {
                                            "type": "integer",
                                            "example": "12"
                                        },
                                        "category_name": {
                                            "type": "string",
                                            "example": "category_name"
                                        },
                                        "created_at": {
                                            "type": "timestamp",
                                            "example": "2021-08-05 01:19:35"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "405": {
                        "description": "Method Not Allowed"
                    },
                    "406": {
                        "description": "Not Acceptable"
                    },
                    "407": {
                        "description": "Proxy Authentication Required"
                    },
                    "410": {
                        "description": "Resource Gone"
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "423": {
                        "description": "Resource Locked"
                    },
                    "429": {
                        "description": "Too Many Requests"
                    },
                    "451": {
                        "description": "Unavailable For Legal Reasons"
                    },
                    "500": {
                        "description": "Internal Server"
                    },
                    "502": {
                        "description": "Bad Gateway"
                    },
                    "503": {
                        "description": "Service Unavailable"
                    },
                    "504": {
                        "description": "Gateway Timeout"
                    },
                    "505": {
                        "description": "HTTP Version Not Supported"
                    },
                    "511": {
                        "description": "Network Authentication Required"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/api/v1/shop/tags": {
            "get": {
                "tags": [
                    "Tag for shops"
                ],
                "summary": "get all tags. => for shopkeeper and system admin",
                "description": "get all tags",
                "operationId": "get categories",
                "responses": {
                    "511": {
                        "description": "Network Authentication Required"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Tag for shops"
                ],
                "summary": "create new tag. => only for system admin",
                "description": "create new tag",
                "operationId": "newTag",
                "responses": {
                    "511": {
                        "description": "Network Authentication Required"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/v1/shop/tags/{id}": {
            "get": {
                "tags": [
                    "Tag for shops"
                ],
                "summary": "get Tag by id in url. => for shopkeeper and system admin",
                "description": "get Tag by id in patch 'url'",
                "operationId": "get tags id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Tag",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "511": {
                        "description": "Network Authentication Required"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Tag for shops"
                ],
                "summary": "update tag. => only for system admin",
                "description": "update tag",
                "operationId": "updateTags",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Category",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "511": {
                        "description": "Network Authentication Required"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Tag for shops"
                ],
                "summary": "delete tag. => only for system admin",
                "description": "delete tag",
                "operationId": "deleteTag",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Tag",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "511": {
                        "description": "Network Authentication Required"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "description": "Authorization token obtained from logging in. Just enter the amount of the token.",
                "name": "Authorization",
                "in": "header",
                "scheme": "Bearer"
            },
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/api/v1/authorize",
                        "tokenUrl": "http://localhost/api/v1/token",
                        "refreshUrl": "http://localhost/api/v1/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    },
...
    "security": [
        []
    ]
}